Proposition of e2e_propagtion.py reshuffling. WIP
